Refactor: separate create domain action saga
