+++ src/modules/dashboard/sagas/actions.ts
@@ -1,6 +1,6 @@
 import { call, fork, put, takeEvery } from 'redux-saga/effects';
 import { hexlify, hexZeroPad } from 'ethers/utils';
-import { ClientType, ROOT_DOMAIN_ID } from '@colony/colony-js';
+import { ClientType } from '@colony/colony-js';
 
 import { ContextModule, TEMP_getContext } from '~context/index';
 import {
@@ -25,168 +25,6 @@ import {
   transactionAddParams,
 } from '../../core/actionCreators';
 import { updateColonyDisplayCache } from './utils';
-
-function* createDomainAction({
-  payload: {
-    colonyAddress,
-    colonyName,
-    domainName,
-    domainColor,
-    domainPurpose,
-    annotationMessage,
-    parentId = ROOT_DOMAIN_ID,
-  },
-  meta: { id: metaId, history },
-  meta,
-}: Action<ActionTypes.COLONY_ACTION_DOMAIN_CREATE>) {
-  let txChannel;
-  try {
-    const apolloClient = TEMP_getContext(ContextModule.ApolloClient);
-    /*
-     * Validate the required values for the payment
-     */
-    if (!domainName) {
-      throw new Error('A domain name is required to create a new domain');
-    }
-
-    txChannel = yield call(getTxChannel, metaId);
-
-    const batchKey = 'createDomainAction';
-    const {
-      createDomainAction: createDomain,
-      annotateCreateDomainAction: annotateCreateDomain,
-    } = yield createTransactionChannels(metaId, [
-      'createDomainAction',
-      'annotateCreateDomainAction',
-    ]);
-
-    const createGroupTransaction = ({ id, index }, config) =>
-      fork(createTransaction, id, {
-        ...config,
-        group: {
-          key: batchKey,
-          id: metaId,
-          index,
-        },
-      });
-
-    yield createGroupTransaction(createDomain, {
-      context: ClientType.ColonyClient,
-      methodName: 'addDomainWithProofs',
-      identifier: colonyAddress,
-      params: [],
-      ready: false,
-    });
-
-    if (annotationMessage) {
-      yield createGroupTransaction(annotateCreateDomain, {
-        context: ClientType.ColonyClient,
-        methodName: 'annotateTransaction',
-        identifier: colonyAddress,
-        params: [],
-        ready: false,
-      });
-    }
-
-    yield takeFrom(createDomain.channel, ActionTypes.TRANSACTION_CREATED);
-    if (annotationMessage) {
-      yield takeFrom(
-        annotateCreateDomain.channel,
-        ActionTypes.TRANSACTION_CREATED,
-      );
-    }
-
-    yield put(transactionPending(createDomain.id));
-
-    /*
-     * Upload domain metadata to IPFS
-     */
-    let domainMetadataIpfsHash = null;
-    domainMetadataIpfsHash = yield call(
-      ipfsUpload,
-      JSON.stringify({
-        domainName,
-        domainColor,
-        domainPurpose,
-      }),
-    );
-
-    yield put(
-      transactionAddParams(createDomain.id, [
-        parentId,
-        (domainMetadataIpfsHash as unknown) as string,
-      ]),
-    );
-
-    yield put(transactionReady(createDomain.id));
-
-    const {
-      payload: { hash: txHash },
-    } = yield takeFrom(
-      createDomain.channel,
-      ActionTypes.TRANSACTION_HASH_RECEIVED,
-    );
-    yield takeFrom(createDomain.channel, ActionTypes.TRANSACTION_SUCCEEDED);
-
-    if (annotationMessage) {
-      yield put(transactionPending(annotateCreateDomain.id));
-
-      /*
-       * Upload domain metadata to IPFS
-       */
-      let annotationMessageIpfsHash = null;
-      annotationMessageIpfsHash = yield call(
-        ipfsUpload,
-        JSON.stringify({
-          annotationMessage,
-        }),
-      );
-
-      yield put(
-        transactionAddParams(annotateCreateDomain.id, [
-          txHash,
-          annotationMessageIpfsHash,
-        ]),
-      );
-
-      yield put(transactionReady(annotateCreateDomain.id));
-
-      yield takeFrom(
-        annotateCreateDomain.channel,
-        ActionTypes.TRANSACTION_SUCCEEDED,
-      );
-    }
-
-    /*
-     * Update the colony object cache
-     */
-    yield apolloClient.query<ColonyFromNameQuery, ColonyFromNameQueryVariables>(
-      {
-        query: ColonyFromNameDocument,
-        variables: { name: colonyName || '', address: colonyAddress },
-        fetchPolicy: 'network-only',
-      },
-    );
-
-    yield put<AllActions>({
-      type: ActionTypes.COLONY_ACTION_DOMAIN_CREATE_SUCCESS,
-      meta,
-    });
-
-    if (colonyName) {
-      yield routeRedirect(`/colony/${colonyName}/tx/${txHash}`, history);
-    }
-  } catch (error) {
-    return yield putError(
-      ActionTypes.COLONY_ACTION_DOMAIN_CREATE_ERROR,
-      error,
-      meta,
-    );
-  } finally {
-    txChannel.close();
-  }
-  return null;
-}
 
 function* editDomainAction({
   payload: {
@@ -751,7 +589,6 @@ function* colonyTokenUnlock({
 }
 
 export default function* tasksSagas() {
-  yield takeEvery(ActionTypes.COLONY_ACTION_DOMAIN_CREATE, createDomainAction);
   yield takeEvery(ActionTypes.COLONY_ACTION_EDIT_COLONY, editColonyAction);
   yield takeEvery(ActionTypes.COLONY_ACTION_DOMAIN_EDIT, editDomainAction);
   yield takeEvery(
